Introduce `gear-voucher`s internal module
