+++ pallets/gear-voucher/Cargo.toml
@@ -18,6 +18,8 @@ scale-info = { workspace = true, features = ["derive"] }
 log.workspace = true
 parity-wasm = { workspace = true, optional = true }
 primitive-types = { workspace = true, features = ["scale-info"] }
+hex = { workspace = true, features = ["alloc"] }
+derive_more.workspace = true
 
 # Internal deps
 common.workspace = true
